tree: replace valstr from libfmap with local version
libfmap is not used for anything fmap anymore, so we should get rid of
the library if possible.

We only have two symbols left from libfmap in our dependencies,
val2str_default and val2str, both of which are trivial functions, so
just copy them to the mosys tree and drop the dependency.

BUG=chromium:990438
TEST=compiles

Change-Id: Ie748d8fb344e6849ebf2439d628e0f0937a69315
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2490580
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>