platform/x86_64/rambi: Switch to 'HOST_FIRMWARE' target
Squash in,

lib/flashrom: Simplify target_bus param passing logic

We wish to remove manual specification of the bus parameter to the
so-called 'internal' flashrom driver. This removes unused cases and
deals with the removal of INTERNAL_BUS_SPI from 'rambi'.

In the case of INTERNAL_BUS_SPI, only 'rambi' boards make use of it
just because the code is old.

Switch flashrom invocations:
 - $ flashrom -p internal:bus=spi [..]
 + $ flashrom -p host             [..]
on rambi devices when fetching the BOOT_STUB section to get spd data.

BUG=b:170914222,b:171603352
TEST=verified on banjo.

Change-Id: Id39c62dca405ce82e5f806b0c6a9515b8e19c0d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2498181
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>